refactor(api): move school security pre-handlers from shared
