+++ api/src/school/domain/usecases/is-school-session-active.js
@@ -1,0 +1,5 @@
+export async function isSchoolSessionActive({ schoolCode, schoolRepository }) {
+  const sessionExpirationDate = await schoolRepository.getSessionExpirationDate({ code: schoolCode });
+
+  return new Date() < new Date(sessionExpirationDate);
+}
